fix: fix logging in iOS.
